+++ codegen/CHANGELOG.md
@@ -1,6 +1,14 @@
 # [exchange-collection-codegen](./README.md) changelog
 > Notable changes to codegen
 
+## [0.2.1] - 2025-02-09
+### Added
+- progress bar
+
+## [0.2.0] - 2025-02-09
+### Modified
+- updated websocket to use the new template
+- force write websocket
 
 ## [0.1.0] - 2024-09-08
 ### Added